removed redunandant and duplicate code, improved naming of variables
