+++ minitwit-api/api/sim/sim-facade
@@ -1,0 +1,47 @@
+package latest
+
+import (
+	"encoding/json"
+	"net/http"
+	"os"
+	"strconv"
+)
+
+func UpdateLatest(r *http.Request) {
+	r.ParseForm()
+	latest := r.Form.Get("latest")
+	if latest != "" {
+		_ = os.WriteFile("./latest_processed_sim_action_id.txt", []byte((latest)), 0644)
+	}
+}
+
+func Get_latest(w http.ResponseWriter, r *http.Request) {
+	content, _ := os.ReadFile("./latest_processed_sim_action_id.txt")
+	latest_command_id, err := strconv.Atoi(string(content))
+	if err != nil {
+		latest_command_id = -1
+	}
+	json.NewEncoder(w).Encode(struct {
+		Latest int `json:"latest"`
+	}{
+		Latest: latest_command_id,
+	})
+}
+
+func Is_authenticated(w http.ResponseWriter, r *http.Request) bool {
+	from_simulator := r.Header.Get("Authorization")
+	if from_simulator != "Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh" {
+		errMsg := "You are not authorized to use this resource!"
+		w.WriteHeader(http.StatusForbidden)
+
+		json.NewEncoder(w).Encode(struct {
+			Status   int    `json:"status"`
+			ErrorMsg string `json:"error_msg"`
+		}{
+			Status:   403,
+			ErrorMsg: errMsg,
+		})
+		return false
+	}
+	return true
+}
