+++ minitwit-api/db/db-facade.go
@@ -3,8 +3,6 @@ package db
 import (
 	"database/sql"
 	"fmt"
-	"net/http"
-	"os"
 
 	"golang.org/x/crypto/bcrypt"
 	"minitwit.com/model"
@@ -12,12 +10,85 @@ import (
 
 const (
 	DATABASE = "./minitwit.db"
-	PER_PAGE = 30
 )
 
 func Connect_db() (db *sql.DB, err error) {
 	fmt.Println("Connecting to database...")
 	return sql.Open("sqlite3", DATABASE)
+}
+
+func GetMessages(args []any, one bool) []map[string]any {
+	query := `SELECT message.*, user.* FROM message, user
+        WHERE message.flagged = 0 AND message.author_id = user.user_id
+        ORDER BY message.pub_date DESC LIMIT ?`
+
+	db, _ := Connect_db()
+	defer db.Close()
+	cur, _ := db.Query(query, args...)
+	defer cur.Close()
+
+	var Messages []map[string]any
+
+	for cur.Next() {
+		var rv model.UserMessageRow
+		_ = cur.Scan(&rv.Message_id, &rv.Author_id, &rv.Text, &rv.Pub_date, &rv.Flagged, &rv.User_id, &rv.Username, &rv.Email, &rv.Pw_hash)
+
+		msg := make(map[string]any)
+		msg["content"] = rv.Text
+		msg["pub_date"] = rv.Pub_date
+		msg["user"] = rv.Username
+
+		Messages = append(Messages, msg)
+	}
+	return Messages
+}
+
+func GetMessagesForUser(args []any, one bool) []map[string]any {
+	query := `SELECT message.*, user.* FROM message, user
+	WHERE message.flagged = 0 AND
+	user.user_id = message.author_id AND user.user_id = ?
+	ORDER BY message.pub_date DESC LIMIT ?`
+
+	db, _ := Connect_db()
+	defer db.Close()
+	cur, _ := db.Query(query, args...)
+	defer cur.Close()
+
+	var Messages []map[string]any
+
+	for cur.Next() {
+		var rv model.UserMessageRow
+		_ = cur.Scan(&rv.Message_id, &rv.Author_id, &rv.Text, &rv.Pub_date, &rv.Flagged, &rv.User_id, &rv.Username, &rv.Email, &rv.Pw_hash)
+
+		msg := make(map[string]any)
+		msg["content"] = rv.Text
+		msg["pub_date"] = rv.Pub_date
+		msg["user"] = rv.Username
+
+		Messages = append(Messages, msg)
+	}
+	return Messages
+}
+
+func GetFollowees(args []any, one bool) []string {
+	query := `SELECT user.username FROM user
+			INNER JOIN follower ON follower.whom_id=user.user_id
+			WHERE follower.who_id=?
+			LIMIT ?`
+
+	db, _ := Connect_db()
+	defer db.Close()
+	cur, _ := db.Query(query, args...)
+	defer cur.Close()
+	var Followees []string
+
+	for cur.Next() {
+		var username string
+		_ = cur.Scan(&username)
+
+		Followees = append(Followees, username)
+	}
+	return Followees
 }
 
 func Query_db(query string, args []any, one bool) (any, error) {
@@ -76,100 +147,15 @@ func Get_user_id(username string) (any, error) {
 	return nil, err
 }
 
-func GetMessages(args []any, one bool) []map[string]any {
-	query := `SELECT message.*, user.* FROM message, user
-        WHERE message.flagged = 0 AND message.author_id = user.user_id
-        ORDER BY message.pub_date DESC LIMIT ?`
-
-	db, _ := Connect_db()
-	defer db.Close()
-	cur, _ := db.Query(query, args...)
-	defer cur.Close()
-
-	var Messages []map[string]any
-
-	// TODO handle empty db
-
-	for cur.Next() {
-		var rv model.UserMessageRow
-		_ = cur.Scan(&rv.Message_id, &rv.Author_id, &rv.Text, &rv.Pub_date, &rv.Flagged, &rv.User_id, &rv.Username, &rv.Email, &rv.Pw_hash)
-
-		msg := make(map[string]any)
-		msg["content"] = rv.Text
-		msg["pub_date"] = rv.Pub_date
-		msg["user"] = rv.Username
-
-		Messages = append(Messages, msg)
-	}
-	return Messages
-}
-
-func GetFollowers(args []any, one bool) []string {
-	query := `SELECT user.username FROM user
-			INNER JOIN follower ON follower.whom_id=user.user_id
-			WHERE follower.who_id=?
-			LIMIT ?`
-
-	db, _ := Connect_db()
-	defer db.Close()
-	cur, _ := db.Query(query, args...)
-	defer cur.Close()
-	var Followers []string
-
-	for cur.Next() {
-		var username string
-		_ = cur.Scan(&username)
-
-		Followers = append(Followers, username)
-	}
-	return Followers
-}
-
-func GetMessagesForUser(args []any, one bool) []map[string]any {
-	query := `SELECT message.*, user.* FROM message, user
-	WHERE message.flagged = 0 AND
-	user.user_id = message.author_id AND user.user_id = ?
-	ORDER BY message.pub_date DESC LIMIT ?`
-
-	db, _ := Connect_db()
-	defer db.Close()
-	cur, _ := db.Query(query, args...)
-	defer cur.Close()
-
-	var Messages []map[string]any
-
-	for cur.Next() {
-		var rv model.UserMessageRow
-		_ = cur.Scan(&rv.Message_id, &rv.Author_id, &rv.Text, &rv.Pub_date, &rv.Flagged, &rv.User_id, &rv.Username, &rv.Email, &rv.Pw_hash)
-
-		msg := make(map[string]any)
-		msg["content"] = rv.Text
-		msg["pub_date"] = rv.Pub_date
-		msg["user"] = rv.Username
-
-		Messages = append(Messages, msg)
-	}
-	return Messages
-}
-
-// ChatGPT
-func IsNil(i interface{}) bool {
-	if i == nil || i == interface{}(nil) {
-		return true
-	} else {
-		return false
-	}
-}
-
 func HashPassword(password string) (string, error) {
 	bytes, err := bcrypt.GenerateFromPassword([]byte(password), 14)
 	return string(bytes), err
 }
 
-func UpdateLatest(r *http.Request) {
-	r.ParseForm()
-	latest := r.Form.Get("latest")
-	if latest != "" {
-		_ = os.WriteFile("./latest_processed_sim_action_id.txt", []byte((latest)), 0644)
+func IsNil(i interface{}) bool {
+	if i == nil || i == interface{}(nil) {
+		return true
+	} else {
+		return false
 	}
 }
