--- minitwit-api/model/Followers_response.go
@@ -1,5 +1,0 @@
-package model
-
-type Followers_response struct {
-	Follows []string
-}
